vector-backend/requirements.txt
# vector-backend/requirements.txt - Updated with Authentication
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2

# OpenAI & Pinecone (optional)
openai==1.3.5
pinecone-client==2.2.4

# Additional utilities
pandas==2.1.3
numpy==1.24.3
tiktoken==0.5.1
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1